feat(hooks): :sparkles: add useItems hook
